Clean up assets and templates
