+++ assets/src/js/revenue-generator-admin.js
@@ -6,5 +6,4 @@ import './pages/revenue-generator-settings';
 import './pages/revenue-generator-welcome';
 import './pages/revenue-generator-paywall';
 import './pages/revenue-generator-dashboard';
-// import './pages/revenue-generator-contribution';
-import './pages/revenue-generator-contribution20';
+import './pages/revenue-generator-contribution';
